+++ python/clu/client.py
@@ -113,14 +113,21 @@ class AMQPClient(BaseClient):
     ----------
     name : str
         The name of the actor.
+    url : str
+        RFC3986 formatted broker address. When used, the other connection
+        keyword arguments are ignored.
     user : str
         The user to connect to the AMQP broker. Defaults to ``guest``.
     password : str
         The password for the user. Defaults to ``guest``.
     host : str
         The host where the AMQP message broker runs. Defaults to ``localhost``.
+    virtualhost : str
+         Virtualhost parameter. ``'/'`` by default.
     port : int
         The port on which the AMQP broker is running. Defaults to 5672.
+    ssl : bool
+        Whether to use TLS/SSL connection.
     version : str
         The version of the actor.
     loop
@@ -143,23 +150,20 @@ class AMQPClient(BaseClient):
 
     connection = None
 
-    def __init__(self, name, user=None, password=None, host=None, port=None,
-                 version=None, loop=None, log_dir=None, log=None,
-                 models=None):
+    def __init__(self, name, url=None, user='guest', password='guest',
+                 host='localhost', port=5672, virtualhost='/', ssl=False,
+                 version=None, loop=None, log_dir=None, log=None, models=None):
 
         super().__init__(name, version=version, loop=loop,
                          log_dir=log_dir, log=log)
 
-        self.user = user or 'guest'
-        self.password = password or 'guest'
-        self.host = host or 'localhost'
-        self.port = port or 5672
-
         self.replies_queue = None
 
         # Creates the connection to the AMQP broker
-        self.connection = TopicListener(self.user, self.password,
-                                        self.host, port=self.port)
+        self.connection = TopicListener(url=url, user=user,
+                                        password=password,
+                                        host=host, port=port,
+                                        ssl=ssl, virtualhost=virtualhost)
 
         #: dict: External commands currently running.
         self.running_commands = {}