+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog
 =========
 
 * :bug:`-` Fix AMQP CLI. It failed when printing a message code ``>`` as HTML.
+* :bug:`44` Add ``url``, ``virtualhost``, and ``ssl`` parameters for `.AMQPClient` that propagates to `.TopicListener`. When defined ``url`` overrides the connection parameters. The CLU CLI now also accepts a ``--url`` flag.
 
 * :release:`0.5.4 <2020-11-05>`
 * :bug:`-` Fix typo that caused `.Device.stop` to fail.